requirements.txt
pytest==8.3.4
pytest-playwright==0.6.2 
allure-pytest==2.13.5